+++ CHANGELOG.md
@@ -10,6 +10,16 @@
 
 ### 4. Hard deprecations
 
+#### Elixir
+
+  * [Date] Deprecate inferring a range with negative step, call `Date.range/3` with a negative step instead
+  * [Enum] Deprecate passing a range with negative step on `Enum.slice/2`, give `first..last//1` instead
+  * [String] Deprecate passing a range with negative step on `String.slice/2`, give `first..last//1` instead
+
+#### ExUnit
+
+  * [ExUnit.Formatter] Deprecate `format_time/2`, use `format_times/1` instead
+
 ## v1.15
 
 The CHANGELOG for v1.15 releases can be found [in the v1.15 branch](https://github.com/elixir-lang/elixir/blob/v1.15/CHANGELOG.md).